requirements.txt
notion-client==2.0.0
openai==1.6.1
httpx>=0.23.0,<0.25.0
python-dotenv==1.0.0
argparse==1.4.0 